Fix helm secret regenerating on upgrade
Whenever a helm upgrade or install is performed, the entire helm chart is
re-evaluated. This causes anything that uses randomly generated values to be
recreated, changing the keys for intial-user-secret. When these secrets changed,
the nginx pod (and any resource that depended on the keys) did not update to
accept the changed keys. This fix applies the random generation only at initial
install time, and preserves the value through helm upgrades.

Signed-off-by: Tudor Marcu <Tudor.Marcu@emc.com>