Remove influxdb pod when cluster is stopped
The influxdb statefulset is completely controlled by the
influxdb-operator, so to remove the pod while keeping the PVC and CRD
in tact we can set size to 0 which will update the statefulset replica
count. Reverting it bring the pod back and the same PVC is used.

Yamllint config updates to make linting less chatty.

Signed-off-by: Tudor Marcu <Tudor.Marcu@emc.com>