[OBSDEF-6350] Add default args for pre-upgrades
The pre-upgrade checks need a `-target-version` argument to know which
version a user is attempting to upgrade to. In the health check screen,
there is no option to choose a version, only a health check to run.
Default to using the current version so that the check can still run
successfully when called without arguments.

Signed-off-by: Tudor Marcu <Tudor.Marcu@emc.com>