[OBSDEF-3870] Make KAHM only set custom values
Currently we build the chart values annotation by transforming the
*entire* .Values file into json, which outputs all settings including
ones that we do not pass in. This makes it impossible to pick out
specific settings and only propagate those during an upgrade.
The default values should be enough to bring up a component, and we only
need to keep track of a small subset that must be kept over an upgrade.
Doing the install this way lets us upgrade KAHM by only changing the
main chart-version annotation, and the rest of the subcharts and values
will be automatically updated from the new chart version.

Signed-off-by: Tudor Marcu <Tudor.Marcu@emc.com>