[OBSDEF-8172] Add fabric proxy rbac for enduser namespaces
Installing an objectstore in an enduser namespace breaks fabric proxy
because it does not have the permissions to get/list pods. We do not
want to enabled the permissions on the default service account (the
default sa attached to init containers), so we need to create our own.
Init containers cannot have separate serviceaccounts than the host Pod,
but we can create a custom API token and mount it as one in the init
container. This lets us keep a separation of privilege between the init
container and other Pod containers.

Signed-off-by: Tudor Marcu <Tudor.Marcu@emc.com>

Fix service account for fabric proxy rbac

Signed-off-by: Tudor Marcu <Tudor.Marcu@emc.com>

Add persistentvolumeclaims for fabricproxy

Signed-off-by: Tudor Marcu <Tudor.Marcu@emc.com>